refactor(frontend): moved download logic to utils
