+++ new/src/frontend/acc2/src/lib/utils.ts
@@ -1,6 +1,19 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 
-export function cn(...inputs: ClassValue[]) {
+export const cn = (...inputs: ClassValue[]): string => {
   return twMerge(clsx(inputs));
-}
+};
+
+export const downloadBlob = (data: Blob, filename: string): void => {
+  const href = URL.createObjectURL(data);
+  const link = document.createElement("a");
+
+  link.href = href;
+  link.download = filename;
+  document.body.appendChild(link);
+  link.click();
+
+  document.body.removeChild(link);
+  URL.revokeObjectURL(href);
+};
